requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fsspec==2025.3.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kornia==0.8.0
kornia_rs==0.1.8
MarkupSafe==3.0.2
modelscope==1.24.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
protobuf==6.30.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
